deslizamientos/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
redis==6.2.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.14.0
uvicorn==0.35.0
zope.interface==7.2
